refactor: Extract iOS E2E test helpers into EditorUITestHelpers
Move reusable helper methods (navigateToEditor, typeInTitle,
insertBlock, typeInContent) and the XCUIElement.waitForEnabled
extension into a dedicated EditorUITestHelpers enum so new test
files can share them without duplication.
