+++ ios/Demo-iOS/GutenbergUITests/EditorInteractionUITest.swift
@@ -24,62 +24,11 @@ final class EditorInteractionUITest: XCTestCase {
         app = nil
     }
 
-    // MARK: - Helpers
-
-    /// Navigates from the editor list through the configuration screen
-    /// and into the full-screen editor. Returns the WebView element once
-    /// the editor has loaded.
-    @discardableResult
-    private func navigateToEditor() throws -> XCUIElement {
-        // Tap the "Default Editor" row in the list.
-        let defaultEditor = app.staticTexts["Default Editor"]
-        XCTAssertTrue(defaultEditor.waitForExistence(timeout: 10), "Default Editor row not found")
-        defaultEditor.tap()
-
-        // Tap the "Start" button on the configuration screen.
-        let startButton = app.buttons["Start"]
-        XCTAssertTrue(startButton.waitForExistence(timeout: 10), "Start button not found")
-        startButton.tap()
-
-        // Wait for the WebView to appear in the full-screen editor.
-        let webView = app.webViews.firstMatch
-        XCTAssertTrue(webView.waitForExistence(timeout: 30), "Expected a WKWebView to appear after editor loads")
-        return webView
-    }
-
-    /// Types text into the title field and returns the field element.
-    @discardableResult
-    private func typeInTitle(_ text: String, webView: XCUIElement) -> XCUIElement {
-        let titleField = webView.textViews["Add title"]
-        XCTAssertTrue(titleField.waitForExistence(timeout: 10), "Title field not found in WebView")
-        titleField.tap()
-        titleField.typeText(text)
-        return titleField
-    }
-
-    /// Opens the block inserter and inserts a block by name.
-    private func insertBlock(_ name: String, webView: XCUIElement) {
-        let addBlockButton = webView.buttons["Add block"]
-        XCTAssertTrue(addBlockButton.waitForExistence(timeout: 10), "Add block button not found in WebView toolbar")
-        addBlockButton.tap()
-
-        let blockOption = app.buttons[name]
-        XCTAssertTrue(blockOption.waitForExistence(timeout: 10), "\(name) block not found in block inserter")
-        blockOption.tap()
-    }
-
-    /// Types text into the currently focused content block.
-    private func typeInContent(_ text: String, webView: XCUIElement) {
-        let block = webView.textViews["Empty block; start writing or type forward slash to choose a block"]
-        XCTAssertTrue(block.waitForExistence(timeout: 10), "Editable block not found in WebView")
-        block.typeText(text)
-    }
-
     // MARK: - Editor Loading
 
     /// A WebView becomes visible after the editor finishes loading.
     func testEditorWebViewBecomesVisible() throws {
-        try navigateToEditor()
+        try EditorUITestHelpers.navigateToEditor(app: app)
     }
 
     // MARK: - Editor History
@@ -94,7 +43,7 @@ final class EditorInteractionUITest: XCTestCase {
     /// 5. Gutenberg JS sends `onEditorHistoryChanged` with `hasRedo: true`
     /// 6. Tap Redo → redo disables, undo re-enables
     func testUndoRedoAfterTyping() throws {
-        let webView = try navigateToEditor()
+        let webView = try EditorUITestHelpers.navigateToEditor(app: app)
 
         let undoButton = app.buttons["Undo"]
         let redoButton = app.buttons["Redo"]
@@ -109,14 +58,14 @@ final class EditorInteractionUITest: XCTestCase {
         XCTAssertFalse(redoButton.isEnabled, "Redo should be disabled on a fresh editor")
 
         // Type in the title field.
-        typeInTitle("Hello", webView: webView)
+        EditorUITestHelpers.typeInTitle("Hello", webView: webView)
 
         // After typing in the title, undo should become enabled.
         XCTAssertTrue(undoButton.waitForEnabled(timeout: 10), "Undo should be enabled after typing in title")
 
         // Insert a Paragraph block and type in it.
-        insertBlock("Paragraph", webView: webView)
-        typeInContent("World", webView: webView)
+        EditorUITestHelpers.insertBlock("Paragraph", webView: webView, app: app)
+        EditorUITestHelpers.typeInContent("World", webView: webView)
 
         // Undo should still be enabled after typing in content.
         XCTAssertTrue(undoButton.isEnabled, "Undo should remain enabled after typing in content")
@@ -139,12 +88,12 @@ final class EditorInteractionUITest: XCTestCase {
     /// verifies the round-trip doesn't crash and the WebView survives
     /// both transitions.
     func testCodeEditorToggleWithContent() throws {
-        let webView = try navigateToEditor()
+        let webView = try EditorUITestHelpers.navigateToEditor(app: app)
 
         // Type content into the title, then insert a Paragraph and type in it.
-        typeInTitle("Test Title", webView: webView)
-        insertBlock("Paragraph", webView: webView)
-        typeInContent("Test content", webView: webView)
+        EditorUITestHelpers.typeInTitle("Test Title", webView: webView)
+        EditorUITestHelpers.insertBlock("Paragraph", webView: webView, app: app)
+        EditorUITestHelpers.typeInContent("Test content", webView: webView)
 
         // Open the overflow menu and switch to Code Editor.
         let moreButton = app.buttons["More"]
@@ -179,9 +128,9 @@ final class EditorInteractionUITest: XCTestCase {
     /// 3. Tap "Image" block → native calls `window.blockInserter.insertBlock()` in JS
     /// 4. Sheet dismisses and block appears in editor
     func testInsertImageBlock() throws {
-        let webView = try navigateToEditor()
+        let webView = try EditorUITestHelpers.navigateToEditor(app: app)
 
-        insertBlock("Image", webView: webView)
+        EditorUITestHelpers.insertBlock("Image", webView: webView, app: app)
 
         // After insertion, an Image block should appear in the editor.
         let imageBlockInEditor = webView.buttons["Upload"]
@@ -189,15 +138,5 @@ final class EditorInteractionUITest: XCTestCase {
             imageBlockInEditor.waitForExistence(timeout: 10),
             "Image block placeholder not found in editor after insertion"
         )
-    }
-}
-
-extension XCUIElement {
-    /// Polls until `isEnabled` becomes `true` or the timeout expires.
-    func waitForEnabled(timeout: TimeInterval) -> Bool {
-        let predicate = NSPredicate(format: "isEnabled == true")
-        let expectation = XCTNSPredicateExpectation(predicate: predicate, object: self)
-        let result = XCTWaiter.wait(for: [expectation], timeout: timeout)
-        return result == .completed
     }
 }
